+++ modules/core/field/src/FarmFieldFactory.php
@@ -146,6 +146,10 @@ class FarmFieldFactory implements FarmFieldFactoryInterface {
         $this->modifyIdTagField($field, $options);
         break;
 
+      case 'inventory':
+        $this->modifyInventoryField($field, $options);
+        break;
+
       case 'list_string':
         $this->modifyListStringField($field, $options);
         break;
@@ -648,6 +652,24 @@ class FarmFieldFactory implements FarmFieldFactoryInterface {
     $field->setDisplayOptions('view', [
       'label' => 'inline',
       'type' => 'id_tag',
+      'weight' => $options['weight']['view'] ?? 0,
+    ]);
+  }
+
+  /**
+   * Inventory field modifier.
+   *
+   * @param \Drupal\Core\Field\BaseFieldDefinition &$field
+   *   A base field definition object.
+   * @param array $options
+   *   An array of options.
+   */
+  protected function modifyInventoryField(BaseFieldDefinition &$field, array $options = []) {
+
+    // Build view display settings.
+    $field->setDisplayOptions('view', [
+      'label' => 'inline',
+      'type' => 'inventory',
       'weight' => $options['weight']['view'] ?? 0,
     ]);
   }