+++ modules/core/inventory/tests/modules/farm_inventory_test/src/Plugin/Asset/AssetType/Container.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Drupal\farm_inventory_test\Plugin\Asset\AssetType;
+
+use Drupal\farm_entity\Plugin\Asset\AssetType\FarmAssetType;
+
+/**
+ * Provides the container asset type.
+ *
+ * @AssetType(
+ *   id = "container",
+ *   label = @Translation("Container"),
+ * )
+ */
+class Container extends FarmAssetType {
+
+}